requirements.txt
allure_python_commons==2.8.1
pytest==6.0.1
requests==2.24.0
PyYAML==5.3.1
